Clean up build process, separate C-library from Perl library installation
It was a big mess because it assumed that both libflux.a and Flux.pm etc
were being installed with the same permissions.  So you had to use sudo
just to compile the libraries, when you should only need to use sudo to
install them.  Added some environment variables FL_PREFIX and PL_PREFIX
to control where the C-library & includes, and the Perl modules get
installed, respectively.  It's not perfect--those variables aren't
preserved across invocations to 'make'.