Improve handling of 3D plotting range parameter.
The parameter $range in simple_relaxer (and passed to
Flux::World::render_lines) was passed by reference, and given a
small margin. This caused successive calls to render a larger
and larger box. Now we sever the dataflow connection so the $range
you pass in to simple_relaxer is the same $range you get out to
pass to subsequent calls.