Fix clang 'address of array is always true' warning
Several places were using a C array variable in a logical statement.
Apparently that just uses the address of the variable, which is
always non-zero, and thus true.  I either checked for the truthiness
of the 0th element of the character array, or just removed the
comparison for a numerical array.