Remove use of $a variable in simple_relaxer.
Using $a (and $b) is problematic because they are not covered
by 'use strict'.  I had changed the $a's to $w's awhile ago,
but apparently forgot one.  This was preventing use of
parallel code for relaxation.

Also changed the documentation in simple_relaxer, to discourage
use of $a = read_world(), etc.