Add a small epsilon to a floating-point comparison in fluxon_auto_open
Without that epsilon, a vertex (say a start or an end vertex) that
was canonically on the open boundary could still be considered to be
on the other side due to a floating-point rounding error.  In one
particularly difficult example, the value was not 1.0 exactly, but
rather 1.0000000000000002.  Beware that just adding DBL_EPSILON to
all the comparisons does not solve all floating-point comparison
problems, but should be OK for this case where the number we're
comparing to is 1.