Fix rgb and prgb in rendering, again.
I had used a reverse slice (-1:0) to make fluxels near the blue
concentrations blue, and fluxels near the red concentrations red.
But that messed up explicit setting of prgb and rgb--they were
backwards (requested yellow was rendered as cyan).  This change
mostly changes the order of alpha and beta in the default rgb and
prgb code, so that these should all be consistent.