Simplify 2 makefile targets. Remove PL_PREFIX from Travis builds.
The return status of 'make pdlinstall' had been the return status
of 'cd ..'.  Not very helpful.  This change makes it return the
status of 'make install'.

Hopefully by removing PL_PREFIX from the Travis builds Flux will
be able to be found, so the install postamble will work correctly, too.